requirements.txt
pyscard==2.0.1
ndeflib==0.3.3